refactor(tsx): move components to their own folder also cleaned up some stuff in d3chart.ts ——— list of assignments → Object.assign (item → accumulator) ——— setTargetAngles commented out & setTargetAnglesNew → setTargetAngles
